spit reordering from selection and add `speak()`
